projects/chat-app/pyproject.toml
[project]
name = "chat-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=1.3.2",
    "openai>=1.55.3",
    "pydantic==2.10.1",
    "python-dotenv>=1.0.1",
]
